Fix for kernel 6.5
The __snd_usbmidi_create function was updated in kernel 6.5 in prep for
MIDI 2.0, which meant we weren't passing enough arguments, resulting
in a gnarly driver crash as the kernel tried to dereference an invalid
address (0x1, presumably garbage left in the argument register).
Update midi.h to match the kernel's version.